Planning: OpenSpace: more fix on compile warning

+++ modules/planning/open_space/hybrid_a_star.cc
@@ -115,16 +115,18 @@ std::shared_ptr<Node3d> HybridAStar::LoadRSPinCS(
 std::shared_ptr<Node3d> HybridAStar::Next_node_generator(
     std::shared_ptr<Node3d> current_node, std::size_t next_node_index) {
   double steering = 0.0;
-  std::size_t index = 0.0;
+  std::size_t index = 0;
   double traveled_distance = 0.0;
-  if (next_node_index < next_node_num_ / 2) {
-    steering = -max_steer_ +
-               (2 * max_steer_ / (next_node_num_ / 2 - 1)) * next_node_index;
+  if (next_node_index < static_cast<double>(next_node_num_) / 2) {
+    steering = -max_steer_ + (2 * max_steer_ /
+                              (static_cast<double>(next_node_num_) / 2 - 1)) *
+                                 static_cast<double>(next_node_index);
     traveled_distance = step_size_;
   } else {
     index = next_node_index - next_node_num_ / 2;
-    steering =
-        -max_steer_ + (2 * max_steer_ / (next_node_num_ / 2 - 1)) * index;
+    steering = -max_steer_ + (2 * max_steer_ /
+                              (static_cast<double>(next_node_num_) / 2 - 1)) *
+                                 static_cast<double>(index);
     traveled_distance = -step_size_;
   }
   // take above motion primitive to generate a curve driving the car to a
